Reshuffle consensus files
