+++ silkworm/core/consensus/base_engine.hpp
@@ -71,17 +71,4 @@ class EngineBase : public IEngine {
                 std::vector<BlockHeader>& old_ommers);
 };
 
-//! \see EIP-1559: Fee market change for ETH 1.0 chain
-std::optional<intx::uint256> expected_base_fee_per_gas(const BlockHeader& parent, const evmc_revision);
-
-//! \see EIP-4844: Shard Blob Transactions
-std::optional<intx::uint256> calc_excess_data_gas(const BlockHeader& parent, std::size_t num_blobs,
-                                                  const evmc_revision);
-
-//! \brief Calculate the transaction root of a block body
-evmc::bytes32 compute_transaction_root(const BlockBody& body);
-
-//! \brief Calculate the hash of ommers of a block body
-evmc::bytes32 compute_ommers_hash(const BlockBody& body);
-
 }  // namespace silkworm::consensus
