+++ silkworm/core/consensus/engine.hpp
@@ -71,15 +71,6 @@ class IEngine {
     virtual evmc::address get_beneficiary(const BlockHeader& header) = 0;
 };
 
-bool transaction_type_is_supported(Transaction::Type, evmc_revision);
-
-//! \brief Performs validation of a transaction that can be done prior to sender recovery and block execution.
-//! \remarks Should sender of transaction not yet recovered a check on signature's validity is performed
-//! \remarks These function is agnostic to whole block validity
-ValidationResult pre_validate_transaction(const Transaction& txn, evmc_revision revision, uint64_t chain_id,
-                                          const std::optional<intx::uint256>& base_fee_per_gas,
-                                          const std::optional<intx::uint256>& data_gas_price);
-
 //! \brief Creates an instance of proper Consensus Engine on behalf of chain configuration
 std::unique_ptr<IEngine> engine_factory(const ChainConfig& chain_config);
