+++ silkworm/core/consensus/ethash_engine.hpp
@@ -18,7 +18,7 @@
 
 #include <ethash/ethash.hpp>
 
-#include <silkworm/core/consensus/base/engine.hpp>
+#include <silkworm/core/consensus/base_engine.hpp>
 
 namespace silkworm::consensus {
