+++ silkworm/core/consensus/pos_engine_test.cpp
@@ -14,11 +14,11 @@
    limitations under the License.
 */
 
-#include "engine.hpp"
+#include "pos_engine.hpp"
 
 #include <catch2/catch.hpp>
 
-#include <silkworm/core/consensus/ethash/engine.hpp>
+#include <silkworm/core/consensus/ethash_engine.hpp>
 #include <silkworm/core/state/in_memory_state.hpp>
 
 namespace silkworm::consensus {
