+++ silkworm/sync/internals/body_sequence.cpp
@@ -17,7 +17,7 @@
 #include "body_sequence.hpp"
 
 #include <silkworm/core/chain/difficulty.hpp>
-#include <silkworm/core/consensus/base/engine.hpp>
+#include <silkworm/core/consensus/validation.hpp>
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/sync/sentry_client.hpp>
